Add files via upload
I did not get to finish all of the assignment. Some of the problems I started on I thought I had finished but then realized I had a constraint I missed and ended up breaking the program trying to meet it. I learned a lot in this assignment, but I know it still takes me a long time to write a program and debug all the issues. Also, it wouldn't let me upload with outputs due to file size.